Fix warnings on x64 windows about size_t to uint conversion in C++ memory allocators.
Fix other 64 bit warnings, sometimes with cast.

+++ storage/falcon/CollationCaseless.cpp
@@ -98,7 +98,7 @@ int CollationCaseless::makeKey(Value *value, IndexKey *key, int partialKey, int 
 	while (q > p && q [-1] == ' ')
 		--q;
 
-	l = q - p;
+	l = (int)(q - p);
 
 	for (int n = 0; n < l; ++n)
 		p [n] = caseTable [p [n]];
